fix: fix func
